perf(phase-5a): refactor SEO barrel imports to explicit subpath imports
17 files updated to import directly from their source modules instead of
re-importing through src/components/seo/index.ts (the SEO barrel). Allows
webpack to tree-shake unused exports per page bundle — previously, importing
BreadcrumbSchema from the barrel would pull in the entire SEO module
(~33 exports including StructuredData, SpeakableSchema, ComparisonSchema,
HinglishKeywords, etc.) before tree-shaking could prune.

Pattern:
- from '@/components/seo'              →  from '@/components/seo/<actual-file>'
- BreadcrumbSchema / COMMON_BREADCRUMBS → @/components/seo/BreadcrumbSchema
- BlogPageWebPageSchema, HowToSchema,
  CoursesPageWebPageSchema, FAQSchema,
  LocationsPageWebPageSchema,
  ResultsPageWebPageSchema             → @/components/seo/StructuredData
- ContentFreshness                     → @/components/seo/ContentFreshness

Barrel index.ts (src/components/seo/index.ts) kept in place for any
future consumers but no production code now uses it. Bundle-analyzer
comparison left for Phase 5d.

Type-check: clean. No DOM, schema, or runtime behavior change.

Refs: AUDIT_FIX_PLAN_2026_05.md (Phase 5a)
