+++ src/app/biology-home-tutor-delhi-ncr/page.tsx
@@ -1,7 +1,7 @@
 import { DelhiAreaSchema } from '@/components/seo/DelhiAreaSchema'
 import type { Metadata } from 'next'
 import Link from 'next/link'
-import { BreadcrumbSchema } from '@/components/seo'
+import { BreadcrumbSchema } from '@/components/seo/BreadcrumbSchema'
 import {
   Phone,
   MessageCircle,
