+++ src/app/blog/[slug]/page.tsx
@@ -2,7 +2,8 @@ import { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { getPostBySlug, getAllPostSlugs, getRelatedPosts, getCategoryBySlug } from '@/lib/blog/mdx'
 import { BlogPostPage } from '@/components/blog/BlogPostPage'
-import { BreadcrumbSchema, COMMON_BREADCRUMBS, HowToSchema } from '@/components/seo'
+import { BreadcrumbSchema, COMMON_BREADCRUMBS } from '@/components/seo/BreadcrumbSchema'
+import { HowToSchema } from '@/components/seo/StructuredData'
 import { TechArticleSchema } from '@/components/seo/TechArticleSchema'
 import { HOWTO_CONFIGS } from '@/data/howto-schemas'
